- Move `test` projects into semantic parent projects - Implement new `:test-utils` project - Resolve requirement for Android SDK installation
