+++ core/build.gradle.kts
@@ -12,7 +12,7 @@ dependencies {
     implementation("org.jsoup:jsoup:1.12.1")
     implementation("com.fasterxml.jackson.module:jackson-module-kotlin:2.11.1")
 
-    testImplementation(project(":testApi"))
+    testImplementation(project(":core:test-api"))
     testImplementation(kotlin("test-junit"))
 }
