+++ core/content-matcher-test-utils/build.gradle.kts
@@ -1,5 +1,5 @@
 dependencies {
-    implementation(project(":testApi"))
+    implementation(project(":core:test-api"))
     implementation(kotlin("stdlib-jdk8"))
     implementation(kotlin("reflect"))
     implementation("com.willowtreeapps.assertk:assertk-jvm:0.22")
