+++ integration-tests/build.gradle.kts
@@ -58,7 +58,7 @@ subprojects {
 
 dependencies {
     implementation(kotlin("stdlib"))
-    implementation(kotlin("test-junit"))
+    api(project(":test-utils"))
     val coroutines_version: String by project
     implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:$coroutines_version")
     implementation("org.jsoup:jsoup:1.12.1")
