+++ plugins/base/base-test-utils/build.gradle.kts
@@ -1,4 +1,4 @@
 dependencies {
     compileOnly(project(":plugins:base"))
-    implementation(project(":testApi"))
+    implementation(project(":core:test-api"))
 }
