+++ plugins/gfm/build.gradle.kts
@@ -3,7 +3,7 @@ import org.jetbrains.registerDokkaArtifactPublication
 dependencies {
     implementation(project(":plugins:base"))
     testImplementation(project(":plugins:base"))
-    testImplementation(project(":plugins:base:test-utils"))
+    testImplementation(project(":plugins:base:base-test-utils"))
 }
 
 registerDokkaArtifactPublication("gfmPlugin") {
