+++ plugins/javadoc/src/test/kotlin/org/jetbrains/dokka/javadoc/JavadocModuleTemplateMapTest.kt
@@ -5,7 +5,7 @@ import org.jetbrains.dokka.javadoc.pages.RowJavadocListEntry
 import org.jetbrains.dokka.links.DRI
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
-import org.jetbrains.dokka.testApi.utils.assertIsInstance
+import org.jetbrains.dokka.test.assertIsInstance
 
 internal class JavadocModuleTemplateMapTest : AbstractJavadocTemplateMapTest() {
