+++ plugins/kotlin-as-java/build.gradle.kts
@@ -3,8 +3,8 @@ import org.jetbrains.registerDokkaArtifactPublication
 dependencies {
     implementation(project(":plugins:base"))
     testImplementation(project(":plugins:base"))
-    testImplementation(project(":plugins:base:test-utils"))
-    testImplementation(project(":test-tools"))
+    testImplementation(project(":plugins:base:base-test-utils"))
+    testImplementation(project(":core:content-matcher-test-utils"))
 }
 
 registerDokkaArtifactPublication("kotlinAsJavaPlugin") {
