+++ runners/gradle-plugin/build.gradle.kts
@@ -19,9 +19,9 @@ dependencies {
     compileOnly("com.android.tools.build:gradle:4.0.1")
     compileOnly(gradleApi())
     compileOnly(gradleKotlinDsl())
+    testImplementation(project(":test-utils"))
     testImplementation(gradleApi())
     testImplementation(gradleKotlinDsl())
-    testImplementation(kotlin("test-junit"))
     testImplementation("org.jetbrains.kotlin:kotlin-gradle-plugin")
     testImplementation("com.android.tools.build:gradle:4.0.1")
