+++ test-utils/build.gradle.kts
@@ -1,0 +1,4 @@
+
+dependencies {
+    api(kotlin("test-junit"))
+}
