+++ test-utils/src/main/kotlin/org/jetbrains/dokka/test/assumeAndroidSdkInstalled.kt
@@ -1,0 +1,8 @@
+package org.jetbrains.dokka.test
+
+import org.junit.Assume.assumeTrue
+
+fun assumeAndroidSdkInstalled() {
+    if (isCI) return
+    assumeTrue(isAndroidSdkInstalled)
+}
